requirements.txt
qrisp[iqm]>=0.7.17
iqm-client>=33.0.1
xgboost>=2.0.0
scikit-learn>=1.3.0
pandas>=2.1.0
numpy>=1.26.0
matplotlib>=3.8.0
plotly>=5.18.0
jupyterlab>=4.0.0
notebook>=7.0.0
scipy>=1.11.0
onnx
